requirements.txt
fastapi 
bcrypt 
pyjwt 
passlib
uvicorn 
SQLAlchemy
pydantic